+++ CHANGES.rst
@@ -3,6 +3,7 @@ Change History
 
 0.2 - Unreleased
 ----------------
+- define `IAutowire`.
 
 0.1.1 - 2018-06-09
 ------------------